fix issues with double states
some states in type-detector are duplicated (for example ON in lights). One of them has no default state (which is good, because it is not shown) but the simple find for some reason tends to find the state without default role, so a role is not assigned and the state is not detected by type-detector anymore and "lost" after saving.

Now added a second find  to make sure that the state with defaultRole is used, if existing.

Not sure if the fallback is needed at all...